:sparkles: Séparation des composants Actions
